nxfx01-api/pyproject.toml
[project]
name = "nxfx01-api"
version = "0.1.0"
description = "NXFX01 Launch Intelligence API — workers, scoring, and FastAPI layer for Base token launch analysis"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "asyncpg>=0.29.0",
    "httpx>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "web3>=6.15.0",
    "cachetools>=5.3.0",
    "websockets>=12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
